requirements.txt
torch
torchvision
numpy
pandas
scikit-learn
pytorch_ignite
pytest==3.10.0
blackhc.laaos~=0.1
blackhc.notebook
blackhc.progress_bar
typing==3.6.6
dataclasses==0.6
pytest-benchmark
